# internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/user/{uid}:
        get:
            tags:
                - User
            operationId: User_GetUser
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.UserItem'
    /v1/user/{uid}/status:
        put:
            tags:
                - User
            operationId: User_UpdateUserStatus
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.UpdateUserStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.UpdateUserStatusReply'
    /v1/users:
        get:
            tags:
                - User
            operationId: User_ListUser
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.ListUserReply'
    /v1/users/select:
        get:
            tags:
                - User
            operationId: User_SelectUser
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lastUID
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.SelectUserReply'
components:
    schemas:
...
        goddess.api.v1.ListUserReply:
            type: object
            properties:
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/goddess.api.v1.UserItem'
...
        goddess.api.v1.SelectUserReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/goddess.api.v1.UserItem'
                total:
                    type: string
                lastUID:
                    type: string
                hasMore:
                    type: boolean
...
        goddess.api.v1.UpdateUserStatusReply:
            type: object
            properties: {}
        goddess.api.v1.UpdateUserStatusRequest:
            type: object
            properties:
                uid:
                    type: string
                status:
                    type: integer
                    format: enum
        goddess.api.v1.UserItem:
            type: object
            properties:
                uid:
                    type: string
                username:
                    type: string
                nickname:
                    type: string
                avatar:
                    type: string
                email:
                    type: string
                phone:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
                status:
                    type: integer
                    format: enum
...